feat: Fix domain models to abstract persistence layer, fix data types (#14)

+++ src/main/java/com/itesm/domain/repository/UserRepository.java
@@ -1,11 +1,14 @@
 package com.itesm.domain.repository;
 
 import com.itesm.domain.models.User;
-import java.util.Optional;
-import java.util.UUID;
 
+import java.util.Optional;
+
+/** UserRepository */
 public interface UserRepository {
     User save(User user);
+
     Optional<User> findByProviderUuid(String providerUuid);
-    Optional<User> findDomainById(Integer id);
+
+    Optional<User> findDomainById(Long id);
 }
